Reapply "Merge pull request #26 from gaabora/main"
This reverts commit c0ad450b4c5edc27c57ef9fb2d375eaa9452d1a4.

+++ reaper_csurf_integrator/oscpkt.hh
@@ -416,7 +416,7 @@ public:
   }
     
   Message &pushStr(const char *s) {
-    int sSize = strlen(s);
+    int sSize = (int) strlen(s);
     assert(sSize < 2147483647); // insane values are not welcome
     type_tags += TYPE_TAG_STRING;
     arguments.push_back(std::make_pair(storage.size(), sSize + 1));
